refactor: move bunwarpJ controller inside the elastic package
