+++ src/main/java/com/ds4h/view/mainGUI/MainMenuGUI.java
@@ -3,7 +3,7 @@ package com.ds4h.view.mainGUI;
 import com.ds4h.controller.alignmentController.AlignmentControllerInterface;
 import com.ds4h.controller.alignmentController.AutomaticAlignmentController.AutomaticAlignmentController;
 import com.ds4h.controller.alignmentController.ManualAlignmentController.ManualAlignmentController;
-import com.ds4h.controller.bunwarpJController.BunwarpJController;
+import com.ds4h.controller.elastic.bunwarpJController.BunwarpJController;
 import com.ds4h.controller.directoryManager.DirectoryManager;
 import com.ds4h.controller.elastic.ElasticController;
 import com.ds4h.controller.elastic.ElasticOutputImageController;
@@ -16,7 +16,6 @@ import com.ds4h.controller.pointController.ImageManagerController;
 import com.ds4h.model.alignment.alignmentAlgorithm.AffineAlignment;
 import com.ds4h.model.alignment.alignmentAlgorithm.ProjectiveAlignment;
 import com.ds4h.model.alignment.alignmentAlgorithm.TranslationalAlignment;
-import com.ds4h.model.alignment.automatic.pointDetector.Detectors;
 import com.ds4h.model.image.imagePoints.ImagePoints;
 import com.ds4h.view.aboutGUI.AboutGUI;
 import com.ds4h.view.automaticAlignmentConfigGUI.AutomaticAlignmentConfigGUI;
